ASSIGNMENT 1
You are given an integer array 'A' of size 'N', sorted in nondecreasing
order. You are also given an integer 'target'. Your task
is to write a function to search for 'target' in the array 'A'. If it
exists, return its index in 0-based indexing using Linear Search in
C++/Python/Java.